blocking mode support.
